Quick bolus: apply code review fixes
- Safety: disable interactive sheet dismissal while bolus enactment is in
  flight (.interactiveDismissDisabled(isEnacting))
- Visual: center pills when only 1–2 suggestions are shown using Spacers
- Correctness: deduplicate suggestions before ForEach to prevent SwiftUI
  identity collisions on repeated Decimal values
- I18n: share single localized title string between .navigationTitle and
  .principal ToolbarItem to prevent translation drift
- A11y: restore minHeight: 44 floor on pill tap target after removing
  minHeight: 64
- Efficiency: capture displayedSuggestions once in pillRow to avoid
  redundant sort per pill
- Convention: move TreatmentsSettingsView to its own file and register
  it in the Xcode project
- Bug fix: reset SlideToConfirmView dragOffset after action fires so the
  thumb returns to start position on auth failure
